requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
google-generativeai
transformers
torch
pysentimiento
python-dotenv
jinja2
aiofiles
uuid
